+++ sdk/include/opentelemetry/sdk/metrics/sync_instruments.h
@@ -51,6 +51,7 @@ public:
   void Add(uint64_t value, const opentelemetry::context::Context &context) noexcept override;
 
 #ifdef OPENTELEMETRY_HAVE_METRICS_BOUND_INSTRUMENTS_PREVIEW
+  using opentelemetry::metrics::Counter<uint64_t>::Bind;
   opentelemetry::nostd::unique_ptr<opentelemetry::metrics::BoundCounter<uint64_t>> Bind(
       const opentelemetry::common::KeyValueIterable &attributes) noexcept override;
 #endif
@@ -73,6 +74,7 @@ public:
   void Add(double value, const opentelemetry::context::Context &context) noexcept override;
 
 #ifdef OPENTELEMETRY_HAVE_METRICS_BOUND_INSTRUMENTS_PREVIEW
+  using opentelemetry::metrics::Counter<double>::Bind;
   opentelemetry::nostd::unique_ptr<opentelemetry::metrics::BoundCounter<double>> Bind(
       const opentelemetry::common::KeyValueIterable &attributes) noexcept override;
 #endif
@@ -92,6 +94,11 @@ public:
 
   void Add(int64_t value) noexcept override;
   void Add(int64_t value, const opentelemetry::context::Context &context) noexcept override;
+#ifdef OPENTELEMETRY_HAVE_METRICS_BOUND_INSTRUMENTS_PREVIEW
+  using opentelemetry::metrics::UpDownCounter<int64_t>::Bind;
+  opentelemetry::nostd::unique_ptr<opentelemetry::metrics::BoundUpDownCounter<int64_t>> Bind(
+      const opentelemetry::common::KeyValueIterable &attributes) noexcept override;
+#endif
 };
 
 class DoubleUpDownCounter : public Synchronous, public opentelemetry::metrics::UpDownCounter<double>
@@ -108,6 +115,11 @@ public:
 
   void Add(double value) noexcept override;
   void Add(double value, const opentelemetry::context::Context &context) noexcept override;
+#ifdef OPENTELEMETRY_HAVE_METRICS_BOUND_INSTRUMENTS_PREVIEW
+  using opentelemetry::metrics::UpDownCounter<double>::Bind;
+  opentelemetry::nostd::unique_ptr<opentelemetry::metrics::BoundUpDownCounter<double>> Bind(
+      const opentelemetry::common::KeyValueIterable &attributes) noexcept override;
+#endif
 };
 
 #if OPENTELEMETRY_ABI_VERSION_NO >= 2
@@ -125,6 +137,11 @@ public:
 
   void Record(int64_t value) noexcept override;
   void Record(int64_t value, const opentelemetry::context::Context &context) noexcept override;
+#  ifdef OPENTELEMETRY_HAVE_METRICS_BOUND_INSTRUMENTS_PREVIEW
+  using opentelemetry::metrics::Gauge<int64_t>::Bind;
+  opentelemetry::nostd::unique_ptr<opentelemetry::metrics::BoundGauge<int64_t>> Bind(
+      const opentelemetry::common::KeyValueIterable &attributes) noexcept override;
+#  endif
 };
 
 class DoubleGauge : public Synchronous, public opentelemetry::metrics::Gauge<double>
@@ -141,6 +158,11 @@ public:
 
   void Record(double value) noexcept override;
   void Record(double value, const opentelemetry::context::Context &context) noexcept override;
+#  ifdef OPENTELEMETRY_HAVE_METRICS_BOUND_INSTRUMENTS_PREVIEW
+  using opentelemetry::metrics::Gauge<double>::Bind;
+  opentelemetry::nostd::unique_ptr<opentelemetry::metrics::BoundGauge<double>> Bind(
+      const opentelemetry::common::KeyValueIterable &attributes) noexcept override;
+#  endif
 };
 #endif
 
@@ -158,6 +180,7 @@ public:
 #endif
 
 #ifdef OPENTELEMETRY_HAVE_METRICS_BOUND_INSTRUMENTS_PREVIEW
+  using opentelemetry::metrics::Histogram<uint64_t>::Bind;
   opentelemetry::nostd::unique_ptr<opentelemetry::metrics::BoundHistogram<uint64_t>> Bind(
       const opentelemetry::common::KeyValueIterable &attributes) noexcept override;
 #endif
@@ -183,6 +206,7 @@ public:
 #endif
 
 #ifdef OPENTELEMETRY_HAVE_METRICS_BOUND_INSTRUMENTS_PREVIEW
+  using opentelemetry::metrics::Histogram<double>::Bind;
   opentelemetry::nostd::unique_ptr<opentelemetry::metrics::BoundHistogram<double>> Bind(
       const opentelemetry::common::KeyValueIterable &attributes) noexcept override;
 #endif